# schemas/openapi.yaml
...
openapi: 3.1.0
info:
  title: SauronID Core API
  version: 0.1.0
  description: |
    HTTP API of the SauronID Rust core (axum).

    Error envelope: 4xx responses from the central error type use
    `{"error":{"code":"snake_case","message":"...","fix":"..."}}` (being
    introduced in a parallel change). Some legacy handlers still return a
    plain-text error string body; both content types are documented on the
    shared error responses.

    Per-call signature (call-sig v2): operations marked with the
    `x-sauron-call-*` header parameters are behind the `require_call_signature`
    middleware. The client signs a canonical, length-prefixed payload under
    domain `sauron.call.v2` with the agent's registered Ed25519 PoP key. Signed
    fields, in order: version, agent_id, tenant_id, audience, method,
    target_uri, content_type, body_sha256, config_digest, timestamp_ms, nonce.
    Enforcement is gated by SAURON_REQUIRE_CALL_SIG (default on in production,
    advisory in development).

    Tenancy: the global tenancy layer extracts `x-sauron-tenant-id` (or the
    admin JWT `tnt` claim) and falls back to the `default` tenant. The header
    is documented on paths whose handlers consume the tenant id.

    Dev endpoints (`/dev/*`) are mounted only when
    SAURON_ENABLE_DEV_ENDPOINTS=1 and are marked `x-dev-only: true`.
servers:
  - url: http://localhost:3001
...
security: []
components:
  securitySchemes:
    adminKey:
      type: apiKey
      in: header
      name: x-admin-key
      description: >-
        Static admin key. Full-write keys allow any method; read-only keys
        allow GET/HEAD only.
    adminJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        HS256 admin JWT. Scopes: admin:read (GET/HEAD), admin:write (mutating),
        admin:super / admin:full / * (super). Optional `tnt` claim restricts
        tenants.
    session:
      type: apiKey
      in: header
      name: x-sauron-session
      description: >-
        Tenant-bound HMAC user session issued by /user/auth or
        /user/auth/finish (format v2|tenant|key_image|exp|sig, 1h TTL).
  parameters:
    TenantId:
      name: x-sauron-tenant-id
      in: header
      required: false
      schema: { type: string, default: default }
      description: >-
        Tenant id consumed by the tenancy layer. Falls back to the `default`
        tenant for legacy callers. Admin JWTs may instead carry a `tnt` claim.
...
  responses:
    BadRequest:
      description: Bad request. Legacy handlers may return a plain-text string instead of the envelope.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        text/plain:
          schema: { type: string }
    Unauthorized:
      description: Unauthorized. Legacy handlers may return a plain-text string instead of the envelope.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        text/plain:
          schema: { type: string }
...
  schemas:
    ErrorEnvelope:
      type: object
      description: >-
        Central 4xx error envelope (being introduced in a parallel change).
        Some legacy handlers still return plain-text bodies.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: snake_case machine-readable code }
            message: { type: string }
            fix: { type: string, description: Actionable remediation hint. }
...
paths:
...
  /v1/attestation/nitro/verify:
    post:
      tags: [v1-attestation]
      summary: Verify an AWS Nitro attestation document
      description: >-
        Validates a raw COSE_Sign1 blob, surfaces the parsed module_id and
        PCRs, and checks the measurement against the operator-registered
        expected hash. Verification failures return 200 with valid=false and
        an error string; only malformed base64 returns 400.
      operationId: nitroVerify
      security: [{ adminKey: [] }, { adminJwt: [] }]
      parameters: [{ $ref: '#/components/parameters/TenantId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [attestation_blob_b64, expected_measurement_hash]
              properties:
                attestation_blob_b64: { type: string, description: base64 AWS Nitro COSE_Sign1 blob. }
                expected_measurement_hash: { type: string, description: Operator-registered measurement (hex). }
                agent_id: { type: [string, "null"], description: Optional; echoed back. }
      responses:
        '200':
          description: Verification result (also used for verification failures, with valid=false).
          content:
            application/json:
              schema:
                type: object
                required: [valid, module_id, pcrs, timestamp, error, agent_id]
                properties:
                  valid: { type: boolean }
                  module_id: { type: string }
                  pcrs:
                    type: object
                    additionalProperties: { type: string }
                    description: PCR index (stringified integer key) to hex SHA-384.
                  timestamp: { type: integer, minimum: 0 }
                  error: { type: [string, "null"], description: Set when valid is false. }
                  agent_id: { type: [string, "null"] }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }